cleaned up repo, reqs, readme
