+++ adult.py
@@ -1,44 +1,44 @@
-import plot
-from data import *
 from plot import *
+from data import *
 from policies import *
-import warnings
-
-warnings.filterwarnings("error", category=RuntimeWarning)
+from pathlib import Path
+import numpy as np
+import plot
 
 
 def main(
-    exp_prefix="trial",
+    exp_suffix="trial",
     group="RAC1P",
     density=1,  # 1 to load the full folktables dataset, smaller than one for a smaller portion.
     n_samples_per_group=5000,  # minimum number of samples for each group: smaller groups are discarded
-    T=2500,
-    n_arms=10,
+    T=5000,
     n_seeds=1,  # number of runs of each policy (same data)
     reg_param=0.01,  # regularization parameter for the ridge regression
     noise_magnitude=0.2,  # None to use the true rewards from the dataset
     expl_coeff_oful=0.01,  # OFUL exploration coefficient. O is equivalent to Greedy.
-    plot_mult=1.5,  # higher value gives a larger plot
+    plot_mult=1.2,  # higher value gives a larger plot
     plot_flag=False,
     epsilon=0.1,
     delta=1e-2,
-    L_tilde=1,
+    L_tilde=None,
+    alpha_regression=None,
+    alpha_delta=1/2,    # defines delta split between regression and relative rank
+    alpha_eps=1/2,      # defines epsilon split between regression and relative rank
     delta_tilde=1e-3,
     noise_type_reg="gaussian",
+    noise_type_rank="zcdp",
+    rescale_bound=None,  # Bound for rescaling features, None for no rescaling
 ):
 
-    exp_dir = (
-        f"exps/adult/{exp_prefix}_na={n_arms}_T={T}_ns={n_seeds}_eps={epsilon}_del={delta}_Lt={L_tilde}_nt={noise_type_reg}/"
-    )
-    run(
-        exp_dir=exp_dir,
+    exp_dir = run(
+        exp_suffix=exp_suffix,
         plot_mult=plot_mult,
         group=group,
         density=density,
-        n_arms=n_arms,
         n_samples_per_group=n_samples_per_group,
         reg_param=reg_param,
         noise_magnitude=noise_magnitude,
+        n_seeds=n_seeds,
         algo_seeds=tuple(range(n_seeds)),
         expl_coeff_oful=expl_coeff_oful,
         T=T,
@@ -47,41 +47,53 @@ def main(
         epsilon=epsilon,
         delta=delta,
         L_tilde=L_tilde,
+        alpha_regression=alpha_regression,
+        alpha_delta=alpha_delta,
+        alpha_eps=alpha_eps,
+        delta_tilde=delta_tilde,
         noise_type_reg=noise_type_reg,
+        noise_type_rank=noise_type_rank,
+        rescale_bound=rescale_bound,
     )
+
     if plot_flag:
         plot.main(dir=f"{exp_dir}plots/", mult=plot_mult, mode_histogram="percentage")
 
+    return exp_dir
+
 
 def run(
+    exp_suffix="trial",
     # problem parameters,
     problem_seed=42,
     mu_noise_level=1e-8,
-    n_arms=10,
     compute_density=False,
     group="RAC1P",
     density=1,
-    n_samples_per_group=5000,
+    n_samples_per_group=50000,
     poly_degree=1,
     noise_magnitude=None,
+    n_seeds=1,
     # algo parameters
     reg_param=0.1,
     expl_coeff_oful=0.1,
     T=5000,  # total number of rounds
-    algo_seeds=tuple(range(10)),
+    algo_seeds=None,
     plot_mult=0.8,
-    exp_dir="exps/adult/simple/",
     plot_flag=False,
     epsilon=0.1,
     delta=1e-2,
-    L_tilde=1,  # bound s.t. ||X||_2^2 + ||y||_2^2 <= L_tilde^2 -- normed to 1 in data.py
+    L_tilde=None,  # bound s.t. ||X||_2^2 + ||y||_2^2 <= L_tilde^2 -- normed to 1 in data.py
     alpha_regression=None,  # confidence parameter --- usual choice is 1/T according to Shariff & Sheffet 2018
+    alpha_delta=1/2,
+    alpha_eps=1/2,
     delta_tilde=1e-3,
     noise_type_reg="gaussian",
+    noise_type_rank="zcdp",
+    rescale_bound=None,  # Bound for rescaling features, None for no rescaling
 ):
-    from pathlib import Path
-
-    Path(f"{exp_dir}plots/").mkdir(parents=True, exist_ok=True)
+    assert T <= n_samples_per_group
+    assert algo_seeds is not None
 
     P = load_adult(
         group=group,
@@ -90,9 +102,25 @@ def run(
         n_samples_per_group=n_samples_per_group,
         seed=problem_seed,
         noise_magnitude=noise_magnitude,
-        n_arms=n_arms,
+        rescale_bound=rescale_bound,
+        L_tilde=L_tilde,
     )
+    L_tilde = P.L_tilde
+    # Dynamically compute L_tilde if needed
+    print(f"[INFO] Retrieved L_tilde: {L_tilde:.4f}...")
+    if L_tilde is None or L_tilde <= 0:
+        raise ValueError("[INFO] L_tilde could not be determined")
+    
     n_arms = P.n_arms
+
+
+    # find trial number manually by checking directories
+    nth_trial = 1
+    exp_dir = f"exps/adult/eps={epsilon}_T={T}_del={delta}_ns={n_seeds}_Lt={L_tilde:.4f}_nt={noise_type_reg}_nr={noise_type_rank}_ad={alpha_delta}_ae={alpha_eps}_{exp_suffix}={nth_trial}/"
+    while Path(exp_dir).exists():
+        nth_trial += 1
+        exp_dir = f"exps/adult/eps={epsilon}_T={T}_del={delta}_ns={n_seeds}_Lt={L_tilde:.4f}_nt={noise_type_reg}_nr={noise_type_rank}_ad={alpha_delta}_ae={alpha_eps}_{exp_suffix}={nth_trial}/"
+    Path(f"{exp_dir}plots/").mkdir(parents=True, exist_ok=True)
 
     params = dict(
         d=P.d,
@@ -108,7 +136,7 @@ def run(
     # plot rewards histograms
     mode_plot_rewards = "density" if compute_density else "hist"
     plot_rewards(
-        P.true_rewards,
+        P.true_rewards.astype(np.float64),
         mode=mode_plot_rewards,
         save=True,
         dir=f"{exp_dir}plots/",
@@ -116,9 +144,8 @@ def run(
         plot_flag=plot_flag,
     )
 
-    # confidence parameter --- usual choice is 1/T according to Shariff & Sheffet 2018
     if alpha_regression is None:
-        alpha_regression = 1
+        alpha_regression = 1/T
     policies_generators = [
         # lambda: Random(),
         lambda: OFUL(reg_param, P.d, expl_coeff_oful),
@@ -129,16 +156,15 @@ def run(
             delta_tilde=delta_tilde,
             L_tilde=L_tilde,
             alpha_regression=alpha_regression,
+            alpha_delta=alpha_delta,
+            alpha_eps=alpha_eps,
             noise_type_reg=noise_type_reg,
+            noise_type_rank=noise_type_rank,
             reg_param=reg_param,
             d=P.d,
             n_arms=n_arms,
         ),
         lambda: FairGreedy(reg_param, P.d, mu_noise_level),
-        # lambda: Greedy(reg_param, P.d),
-        # lambda: FairGreedyKnownCDF(reg_param, P.d, mu_noise_level, P),
-        # lambda: FairGreedyKnownMuStar(P)
-        # lambda: FairGreedyNoNoise(reg_param, P.d),
     ]
 
     total_ps, total_dfs = [], []
@@ -157,6 +183,12 @@ def run(
         policy_name = dfs[0]["policy"].drop_duplicates()[0]
 
         print(f"Results for {policy_name}")
+        for df in dfs:
+            df.hist(
+                column="actions", bins=n_arms,
+            )
+            break
+        
         if plot_flag:
             for df in dfs:
                 df.hist(column="sel_group", bins=P.n_groups)
@@ -169,6 +201,7 @@ def run(
                     df[m].plot()
                 plt.title(f"{m}_{policy_name}")
                 plt.show()
+
         for p in ps:
             print(f"mu_est_MSE = {np.mean((p.get_mu_estimate() - P.mu_star) ** 2)}")
             print(f"mu_est = {p.get_mu_estimate()}")
@@ -188,7 +221,6 @@ def run(
     return exp_dir
 
 
-# NOTE: THIS IS WHERE EXPERIMENTS START
 if __name__ == "__main__":
     import argparse
 
@@ -197,16 +229,32 @@ if __name__ == "__main__":
         "--plot", action="store_true", help="Enable plotting (default: disabled)"
     )
     args = parser.parse_args()
-    for n_arms in (10,):
-        for n_samples_per_group in (5000,):
-            main(
-                n_arms=n_arms,
-                n_samples_per_group=n_samples_per_group,
-                n_seeds=1,
-                plot_flag=args.plot,
-                T=50000,
-                epsilon=15,
-                delta=0.1,
-                L_tilde=1,  # TODO: this needs to be max row norm of X+Y
-                delta_tilde=0.01,
-            )
+
+    exp_dirs = []
+    n_samples_per_groups = (50001,)
+    epsilons = (30,)
+    alpha_delta_epsilons = ((0.9, 0.9),)
+    for n_samples_per_group in n_samples_per_groups:
+        for epsilon in epsilons:
+            for (alpha_delta, alpha_eps) in alpha_delta_epsilons:
+                exp_dir = main(
+                    n_samples_per_group=n_samples_per_group,
+                    n_seeds=1,
+                    plot_flag=args.plot,
+                    T=5000,
+                    epsilon=epsilon,         # total epsilon budget for DPFairGreedy
+                    delta=0.1,          # total delta budget for DPFairGreedy
+                    L_tilde=None,       # max row norm of X+Y, will be computed based on data
+                    alpha_delta=alpha_delta,    # defines delta split between regression and relative rank
+                    alpha_eps=alpha_eps,      # defines epsilon split between regression and relative rank
+                    delta_tilde=0.001,   # slack on the relative rank delta for advanced composition
+                    noise_type_rank="zcdp",
+                    rescale_bound=None,
+                )
+                exp_dirs.append(exp_dir)
+
+    print("\n-------------------------- OUTPUT DIRS BY VARIABLES --------------------------\n")
+    for n_samples_per_group in n_samples_per_groups:
+        for epsilon in epsilons:
+            for (alpha_delta, alpha_eps) in alpha_delta_epsilons:
+                print(f"Output for n_samples={n_samples_per_group}, epsilon={epsilon}, (alpha_delta, alpha_eps)=({alpha_delta}, {alpha_eps}): \n{exp_dir}\n")
