+++ requirements.txt
@@ -1,11 +1,9 @@
 # Generated with pipreqs executed on the machine used to perform the experiments.
 
-folktables==0.0.11
-numpy==1.23.3
-pandas==1.5.0
-scikit-learn==1.1.2
+folktables==0.0.12
+matplotlib==3.5.1
+numpy==1.21.4
+pandas==1.3.5
+scikit_learn==1.0.2
+seaborn==0.13.2
 tqdm==4.62.3
-seaborn==0.12.0
-matplotlib==3.6.0
-
-
